pyproject.toml
[tool.poetry]
name = "brave-leo-with-search"
version = "0.1.0"
description = ""
authors = ["Sneh Pushp <pushp1999satyam@gmail.com>"]
readme = "README.md"

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  # The following are specific to Black, you probably don't want those.
  | blib2to3
  | tests/data
  | profiling
)/
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 120

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.1"
uvicorn = "^0.22.0"
SQLAlchemy = {extras = ["asyncio"], version = "^2.0.13"}
python-multipart = "^0.0.6"
requests = "2.28.2"
pydantic = {version = "<2.1.1", extras = ["email"]}
loguru = "^0.7.2"
langchain = "0.1.20"
langchain-groq = "^0.1.3"
python-dotenv = "^1.0.1"
firecrawl-py = "^0.0.20"
langchain-community = "0.0.38"
tavily-python = "^0.3.3"
langchainhub = "^0.1.20"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
